+++ src/main/java/eda/domain/FeatureViewRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface FeatureViewRepository extends JpaRepository<FeatureView, Integer> {
+public interface FeatureViewRepository extends JpaRepository<FeatureView, Long> {
     Optional<FeatureView> findByName(String name);
 }
